% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRhythms_dodr.R
\name{compareRhythms_dodr}
\alias{compareRhythms_dodr}
\title{Run differential rhythmicity analysis defined in Thaben & Westermark}
\usage{
compareRhythms_dodr(
  expr,
  exp_design,
  period = 24,
  rhythm_fdr = 0.05,
  compare_fdr = 0.05,
  amp_cutoff = 0.5,
  just_classify = TRUE
)
}
\arguments{
\item{expr}{A matrix of expression values with gene in the rows and samples in columns}

\item{exp_design}{A data.frame of the experimental design with at least two
columns: "time" and "group".}

\item{period}{The period of rhythm being tested (default = 24)}

\item{rhythm_fdr}{The false discovery cutoff for finding rhythmic time series
(default = 0.05)}

\item{compare_fdr}{The false discovery cutoff for the comparison of rhythms
(default = 0.05)}

\item{amp_cutoff}{The minimum peak-to-trough amplitude in log2 scale
considered biologically relevant (default = 0.5)}

\item{just_classify}{Boolean specifying whether genes must only be classified
(TRUE) or if the amplitude and phases of fits should also be returned
(FALSE)}
}
\description{
Run differential rhythmicity analysis defined in Thaben & Westermark
}
\keyword{internal}
