% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRhythms_edgeR.R
\name{compareRhythms_edgeR}
\alias{compareRhythms_edgeR}
\title{Run differential rhythmicity analysis for RNA-seq data using edgeR}
\usage{
compareRhythms_edgeR(
  counts,
  exp_design,
  lengths,
  period,
  rhythm_fdr,
  compare_fdr,
  amp_cutoff,
  just_classify
)
}
\arguments{
\item{exp_design}{A data.frame of the experimental design with at least two
columns: "time" and "group".}

\item{lengths}{A data.frame of average transcript lengths. Only used with
methods "deseq" and "edgeR".}

\item{period}{The period of rhythm being tested (default = 24)}

\item{rhythm_fdr}{The false discovery cutoff for finding rhythmic time series
(default = 0.05)}

\item{compare_fdr}{The false discovery cutoff for the comparison of rhythms
(default = 0.05)}

\item{amp_cutoff}{The minimum peak-to-trough amplitude in log2 scale
considered biologically relevant (default = 0.5)}

\item{just_classify}{Boolean specifying whether genes must only be classified
(TRUE) or if the amplitude and phases of fits should also be returned
(FALSE)}
}
\description{
Run differential rhythmicity analysis for RNA-seq data using edgeR
}
\keyword{internal}
