% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRhythms_limma.R
\name{compareRhythms_limma}
\alias{compareRhythms_limma}
\title{Run differential rhythmicity analysis for microarray using limma}
\usage{
compareRhythms_limma(
  eset,
  exp_design,
  period,
  rhythm_fdr,
  compare_fdr,
  amp_cutoff,
  just_classify,
  robust,
  rna_seq
)
}
\arguments{
\item{eset}{A matrix of expression values with gene in the rows and samples in columns}

\item{exp_design}{A data.frame of the experimental design with at least two
columns: "time" and "group".}

\item{period}{The period of rhythm being tested (default = 24)}

\item{rhythm_fdr}{The false discovery cutoff for finding rhythmic time series
(default = 0.05)}

\item{compare_fdr}{The false discovery cutoff for the comparison of rhythms
(default = 0.05)}

\item{amp_cutoff}{The minimum peak-to-trough amplitude in log2 scale
considered biologically relevant (default = 0.5)}

\item{just_classify}{Boolean specifying whether genes must only be classified
(TRUE) or if the amplitude and phases of fits should also be returned
(FALSE)}

\item{robust}{Boolean to turn on robust computation of statistics in
different methods (default = TRUE).}
}
\description{
Run differential rhythmicity analysis for microarray using limma
}
\keyword{internal}
