% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRhythms_model_select.R
\name{compareRhythms_model_select}
\alias{compareRhythms_model_select}
\title{Run differential rhythmicity analysis using linear model selection}
\usage{
compareRhythms_model_select(
  data,
  exp_design,
  period,
  amp_cutoff,
  criterion,
  schwarz_wt_cutoff,
  just_classify
)
}
\arguments{
\item{data}{A matrix of log2 expression values (if microarray), expression
counts (RNA-seq) or normalized data (see Details).}

\item{exp_design}{A data.frame of the experimental design with at least two
columns: "time" and "group".}

\item{period}{The period of rhythm being tested (default = 24)}

\item{amp_cutoff}{The minimum peak-to-trough amplitude in log2 scale
considered biologically relevant (default = 0.5)}

\item{criterion}{The criterion used for model selection. These can be "aic"
or "bic" (default = "bic"). Only used for method = "mod_sel".}

\item{schwarz_wt_cutoff}{The conditional probability that the best model is
the true model. Genes with a conditional probability smaller than this
cutoff are deemed unclassifiable. This is only used for method = "mod_sel".
(default = 0.4)}

\item{just_classify}{Boolean specifying whether genes must only be classified
(TRUE) or if the amplitude and phases of fits should also be returned
(FALSE)}
}
\description{
This function runs the information criteria-based model selection proposed by
Atger et al. (2015) for identifying timeseries with different rhythms in the
two datasets.
}
\keyword{internal}
