% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRhythms_voom.R
\name{compareRhythms_voom}
\alias{compareRhythms_voom}
\title{Run differential rhythmicity analysis for RNA-seq data using limma-voom}
\usage{
compareRhythms_voom(
  counts,
  exp_design,
  period,
  rhythm_fdr,
  compare_fdr,
  amp_cutoff,
  just_classify,
  robust,
  outliers
)
}
\arguments{
\item{counts}{A matrix of  expression counts}

\item{exp_design}{A data.frame of the experimental design with at least two
columns: "time" and "group".}

\item{period}{The period of rhythm being tested (default = 24)}

\item{rhythm_fdr}{The false discovery cutoff for finding rhythmic time series
(default = 0.05)}

\item{compare_fdr}{The false discovery cutoff for the comparison of rhythms
(default = 0.05)}

\item{amp_cutoff}{The minimum peak-to-trough amplitude in log2 scale
considered biologically relevant (default = 0.5)}

\item{just_classify}{Boolean specifying whether genes must only be classified
(TRUE) or if the amplitude and phases of fits should also be returned
(FALSE)}

\item{robust}{Boolean to turn on robust computation of statistics in
different methods (default = TRUE).}

\item{outliers}{Boolean specifying if weights must be computed for each
sample to account for outliers. Only used by method = "voom".}
}
\description{
Run differential rhythmicity analysis for RNA-seq data using limma-voom
}
\keyword{internal}
