% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_fat_diet_ma}
\alias{high_fat_diet_ma}
\title{Effect of high fat diet on the liver transcriptome - Microarray}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 23060 rows and 36 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52333}
}
\usage{
high_fat_diet_ma
}
\description{
Time-series data of the mouse liver transcriptome measured under normal chow
read using custom CDF file from Brainarray version 24.0.0 annotated by
Ensembl Gene ID followed by RMA normalization. The expression data are in the
log2 scale. The sample specification are present in the column names in a
\verb{<group>}_ZT\verb{<time>} format.
}
\keyword{datasets}
