% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_fat_diet_rnaseq}
\alias{high_fat_diet_rnaseq}
\title{Effect of high fat diet on the liver transcriptome - RNA-sequencing}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 37310 rows and 36 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE108688}
}
\usage{
high_fat_diet_rnaseq
}
\description{
Time-series data of the mouse liver transcriptome measured under normal chow
(NC) and high fat diet (HFD). The RNA-seq data aligned using STAR to mm9
genome and quantified using featureCounts (see GSE108688 for details) and
annotated by Ensembl Gene ID. The sample specification are present in the
column names in a \verb{<group>}\emph{ZT\verb{<time>}}\verb{<replicate>} format.
}
\keyword{datasets}
